Splitted ItemsVariations from ItemController into seperate controller, added API routes for Tax
